KAFKA-12435: Fix javadoc errors
