+++ clients/src/main/java/org/apache/kafka/server/log/remote/storage/RemoteLogMetadataManager.java
@@ -71,7 +71,7 @@ public interface RemoteLogMetadataManager extends Configurable, Closeable {
      * <p>
      * <pre>
      * +---------------------+            +----------------------+
-     * |COPY_SEGMENT_STARTED |----------->|COPY_SEGMENT_FINISHED |
+     * |COPY_SEGMENT_STARTED |-----------&gt;|COPY_SEGMENT_FINISHED |
      * +-------------------+-+            +--+-------------------+
      *                     |                 |
      *                     |                 |
