+++ clients/src/main/java/org/apache/kafka/server/log/remote/storage/RemoteLogSegmentState.java
@@ -32,7 +32,7 @@ import java.util.stream.Collectors;
  * <p>
  * <pre>
  * +---------------------+            +----------------------+
- * |COPY_SEGMENT_STARTED |----------->|COPY_SEGMENT_FINISHED |
+ * |COPY_SEGMENT_STARTED |-----------&gt;|COPY_SEGMENT_FINISHED |
  * +-------------------+-+            +--+-------------------+
  *                     |                 |
  *                     |                 |
